+++ lib/cannery_web/live/ammo_group_live/form_component.ex
@@ -9,6 +9,8 @@ defmodule CanneryWeb.AmmoGroupLive.FormComponent do
   alias Ecto.Changeset
   alias Phoenix.LiveView.Socket
 
+  @ammo_group_create_limit 10_000
+
   @impl true
   @spec update(
           %{:ammo_group => AmmoGroup.t(), :current_user => User.t(), optional(any) => any},
@@ -22,6 +24,7 @@ defmodule CanneryWeb.AmmoGroupLive.FormComponent do
   def update(%{assigns: %{ammo_group: ammo_group, current_user: current_user}} = socket) do
     socket =
       socket
+      |> assign(:ammo_group_create_limit, @ammo_group_create_limit)
       |> assign(:changeset, Ammo.change_ammo_group(ammo_group))
       |> assign(:ammo_types, Ammo.list_ammo_types(current_user))
       |> assign_new(:containers, fn -> Containers.list_containers(current_user) end)
@@ -80,20 +83,68 @@ defmodule CanneryWeb.AmmoGroupLive.FormComponent do
   end
 
   defp save_ammo_group(
-         %{assigns: %{current_user: current_user, return_to: return_to}} = socket,
+         %{assigns: %{changeset: changeset}} = socket,
          :new,
-         ammo_group_params
+         %{"multiplier" => multiplier_str} = ammo_group_params
        ) do
     socket =
-      case Ammo.create_ammo_group(ammo_group_params, current_user) do
-        {:ok, _ammo_group} ->
-          prompt = dgettext("prompts", "Ammo group created successfully")
-          socket |> put_flash(:info, prompt) |> push_redirect(to: return_to)
+      case multiplier_str |> Integer.parse() do
+        {multiplier, _remainder}
+        when multiplier >= 1 and multiplier <= @ammo_group_create_limit ->
+          socket |> create_multiple(ammo_group_params, multiplier)
 
-        {:error, %Changeset{} = changeset} ->
-          socket |> assign(changeset: changeset)
+        {multiplier, _remainder} ->
+          error_msg =
+            dgettext(
+              "errors",
+              "Invalid number of copies, must be between 1 and %{max}. Was %{multiplier}",
+              max: @ammo_group_create_limit,
+              multiplier: multiplier
+            )
+
+          {:error, changeset} =
+            changeset
+            |> Changeset.add_error(:multiplier, error_msg)
+            |> Changeset.apply_action(:insert)
+
+          socket |> assign(:changeset, changeset)
+
+        :error ->
+          error_msg = dgettext("errors", "Could not parse number of copies")
+
+          {:error, changeset} =
+            changeset
+            |> Changeset.add_error(:multiplier, error_msg)
+            |> Changeset.apply_action(:insert)
+
+          socket |> assign(:changeset, changeset)
       end
 
     {:noreply, socket}
+  end
+
+  defp create_multiple(
+         %{assigns: %{current_user: current_user, return_to: return_to}} = socket,
+         ammo_group_params,
+         multiplier
+       ) do
+    case Ammo.create_ammo_groups(ammo_group_params, multiplier, current_user) do
+      {:ok, {count, _ammo_groups}} ->
+        prompt =
+          dngettext(
+            "prompts",
+            "Ammo group created successfully",
+            "Ammo groups created successfully",
+            count
+          )
+
+        socket |> put_flash(:info, prompt) |> push_redirect(to: return_to)
+
+      {:error, %Changeset{} = changeset} ->
+        socket |> assign(changeset: changeset)
+
+      {:error, nil} ->
+        socket
+    end
   end
 end