+++ test/cannery_web/live/range_live_test.exs
@@ -16,7 +16,9 @@ defmodule CanneryWeb.RangeLiveTest do
   defp create_shot_group(%{current_user: current_user}) do
     container = container_fixture(%{"staged" => true}, current_user)
     ammo_type = ammo_type_fixture(current_user)
-    ammo_group = ammo_group_fixture(%{"staged" => true}, ammo_type, container, current_user)
+
+    {1, [ammo_group]} =
+      ammo_group_fixture(%{"staged" => true}, ammo_type, container, current_user)
 
     shot_group =
       %{"count" => 5, "date" => ~N[2022-02-13 03:17:00], "notes" => "some notes"}